saving files in directories by user name

+++ bot/parser/photo.py
@@ -11,13 +11,21 @@ class Photo(Message):
     def message(self, message):
         assert isinstance(message.photo, list)
         photo = message.photo[-1]
+        user_id = str(message.from_user.id)
         file_id = photo.file_id
         file = self.bot.get_file(file_id)
-        full_file_path = os.path.join(self.ROOT_FOLDER, file.file_path)
-        os.makedirs(os.path.dirname(full_file_path), exist_ok=True)
-        with open(full_file_path, 'wb') as f:
-            f.write(self.bot.download_file(file.file_path))
-        self.bot.send_message(message.from_user.id,
-                              'Photo Size: {}x{} Total: {}'.format(photo.height, photo.width, photo.file_size))
-
-
+        photo_path = user_id + '/' + file.file_path
+        full_file_path = os.path.join(self.ROOT_FOLDER, photo_path)
+        if os.path.exists(full_file_path):
+            with open(full_file_path, 'wb') as f:
+                f.write(self.bot.download_file(file.file_path))
+            self.bot.send_message(message.from_user.id,
+                                  'Photo Size: {}x{} Total: {} KB'.format(photo.height,
+                                                                          photo.width, photo.file_size//1024))
+        else:
+            os.makedirs(os.path.dirname(full_file_path), exist_ok=True)
+            with open(full_file_path, 'wb') as f:
+                f.write(self.bot.download_file(file.file_path))
+            self.bot.send_message(message.from_user.id,
+                                  'Photo Size: {}x{} Total: {} KB'.format(photo.height,
+                                                                          photo.width, photo.file_size//1024))
